moving tests to one class
